Refactor search functionality to utilize centralized search utilities, improving code organization and maintainability
